lin,cleanup: compiler warnings

+++ src/transform/Linearizer.cpp
@@ -223,7 +223,7 @@ Linearizer::promoteDefinitionExt(SmallVector<Value*, 16> & defs, Value & inst, V
 
   if (defBlockId == destBlockId) return inst;
 
-  const int span = destBlockId - defBlockId;
+  const size_t span = destBlockId - defBlockId;
 
   IF_DEBUG_LIN { errs() << "\t* promoting value " << inst << " from def block " << defBlockId << " to " << destBlockId << "\n"; }
 
@@ -233,7 +233,7 @@ Linearizer::promoteDefinitionExt(SmallVector<Value*, 16> & defs, Value & inst, V
 
   auto instShape = vecInfo.getVectorShape(inst);
 
-  for (int i = 1; i < span + 1; ++i) {
+  for (size_t i = 1; i < span + 1; ++i) {
     int blockId = defBlockId + i;
 
     auto & block = getBlock(blockId);
@@ -943,7 +943,7 @@ Linearizer::createSuperInput(PHINode & phi, SuperInput & superInput) {
   auto & phiBlock = *phi.getParent();
 
   auto phiShape = vecInfo.getVectorShape(phi);
-  for (int i = 1; i < blocks.size(); ++i) {
+  for (size_t i = 1; i < blocks.size(); ++i) {
     auto * inBlock = blocks[i];
     auto * inVal = phi.getIncomingValueForBlock(inBlock);
 
@@ -1039,7 +1039,7 @@ Linearizer::foldPhis(BasicBlock & block) {
     // all inputs that are incoming on this edge after folding
     SmallVector<BasicBlock*, 4> superposedInBlocks;
 
-    for (int i = 0; i < phi.getNumIncomingValues(); ++i) {
+    for (size_t i = 0; i < phi.getNumIncomingValues(); ++i) {
       auto * inBlock = phi.getIncomingBlock(i);
 
       // this incoming block remains an immediate predecessor so its value can only be live in on that block
@@ -1686,7 +1686,7 @@ Linearizer::cacheMasks(){
 
 // cache branch masks
    auto & term = *block.getTerminator();
-   for (int i = 0; i < term.getNumSuccessors(); ++i) {
+   for (size_t i = 0; i < term.getNumSuccessors(); ++i) {
      auto * succBlock = term.getSuccessor(i);
      auto * edgeMask = maskAnalysis.getExitMask(block, *succBlock);
      if (edgeMask) setEdgeMask(block, *succBlock, edgeMask);
@@ -1751,7 +1751,7 @@ Linearizer::fixSSA() {
 
       // phi def/use repair
       if (phi) {
-        for (int inIdx = 0; inIdx < phi->getNumIncomingValues(); ++inIdx) {
+        for (size_t inIdx = 0; inIdx < phi->getNumIncomingValues(); ++inIdx) {
           auto * inBlock = phi->getIncomingBlock(inIdx);
           auto * inVal = phi->getIncomingValue(inIdx);
 
@@ -1775,7 +1775,7 @@ Linearizer::fixSSA() {
       }
 
       // non-phi def/use repair
-      for (int opIdx = 0; opIdx < inst.getNumOperands(); ++opIdx) {
+      for (size_t opIdx = 0; opIdx < inst.getNumOperands(); ++opIdx) {
         auto * opInst = dyn_cast<Instruction>(inst.getOperand(opIdx));
         if (!opInst) continue;
